[+BUGFIX]: explicitly cast offset to integer to avoid exceptions in paginate view helper
This fixes: #10902

+++ typo3/sysext/fluid/Classes/ViewHelpers/Widget/Controller/PaginateController.php
@@ -72,7 +72,7 @@ class Tx_Fluid_ViewHelpers_Widget_Controller_PaginateController extends Tx_Fluid
 		$query = $this->objects->getQuery();
 		$query->setLimit($itemsPerPage);
 		if ($this->currentPage > 1) {
-			$query->setOffset($itemsPerPage * ($this->currentPage - 1));
+			$query->setOffset((integer)($itemsPerPage * ($this->currentPage - 1)));
 		}
 		$modifiedObjects = $query->execute();
